api/requirements.txt
# Backend
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6

# Desarrollo
pytest==7.4.3
pytest-asyncio==0.21.1